+++ keba_keenergy_api/constants.py
@@ -119,7 +119,7 @@ class HeatCircuitExternalHeatRequest(str, Enum):
     ON = "true"
 
 
-PAYLOAD_PREFIX: Final[str] = "APPL.CtrlAppl.sParam"
+PAYLOAD_PREFIX: Final[str] = "APPL.CtrlAppl"
 
 
 class EndpointProperties(NamedTuple):
@@ -135,25 +135,25 @@ class System(Enum):
     """The system endpoint settings."""
 
     HOT_WATER_TANK_NUMBERS = EndpointProperties(
-        f"{PAYLOAD_PREFIX}.options.systemNumberOfHotWaterTanks",
+        f"{PAYLOAD_PREFIX}.sParam.options.systemNumberOfHotWaterTanks",
         value_type=int,
     )
     HEAT_PUMP_NUMBERS = EndpointProperties(
-        f"{PAYLOAD_PREFIX}.options.systemNumberOfHeatPumps",
+        f"{PAYLOAD_PREFIX}.sParam.options.systemNumberOfHeatPumps",
         value_type=int,
     )
     HEAT_CIRCUIT_NUMBERS = EndpointProperties(
-        f"{PAYLOAD_PREFIX}.options.systemNumberOfHeatingCircuits",
+        f"{PAYLOAD_PREFIX}.sParam.options.systemNumberOfHeatingCircuits",
         value_type=int,
     )
     OPERATING_MODE = EndpointProperties(
-        f"{PAYLOAD_PREFIX}.param.operatingMode",
+        f"{PAYLOAD_PREFIX}.sParam.param.operatingMode",
         value_type=int,
         read_only=False,
         human_readable=SystemOperatingMode,
     )
     OUTDOOR_TEMPERATURE = EndpointProperties(
-        f"{PAYLOAD_PREFIX}.outdoorTemp.values.actValue",
+        f"{PAYLOAD_PREFIX}.sParam.outdoorTemp.values.actValue",
         value_type=float,
     )
 
@@ -162,27 +162,27 @@ class HotWaterTank(Enum):
     """The hot water tank endpoint settings."""
 
     TEMPERATURE = EndpointProperties(
-        f"{PAYLOAD_PREFIX}.hotWaterTank[%s].topTemp.values.actValue",
+        f"{PAYLOAD_PREFIX}.sParam.hotWaterTank[%s].topTemp.values.actValue",
         value_type=float,
     )
     OPERATING_MODE = EndpointProperties(
-        f"{PAYLOAD_PREFIX}.hotWaterTank[%s].param.operatingMode",
+        f"{PAYLOAD_PREFIX}.sParam.hotWaterTank[%s].param.operatingMode",
         value_type=int,
         read_only=False,
         human_readable=HotWaterTankOperatingMode,
     )
     MIN_TEMPERATURE = EndpointProperties(
-        f"{PAYLOAD_PREFIX}.hotWaterTank[%s].param.reducedSetTempMax.value",
+        f"{PAYLOAD_PREFIX}.sParam.hotWaterTank[%s].param.reducedSetTempMax.value",
         float,
         read_only=False,
     )
     MAX_TEMPERATURE = EndpointProperties(
-        f"{PAYLOAD_PREFIX}.hotWaterTank[%s].param.normalSetTempMax.value",
+        f"{PAYLOAD_PREFIX}.sParam.hotWaterTank[%s].param.normalSetTempMax.value",
         float,
         read_only=False,
     )
     HEAT_REQUEST = EndpointProperties(
-        f"{PAYLOAD_PREFIX}.hotWaterTank[%s].values.heatRequestTop",
+        f"{PAYLOAD_PREFIX}.sParam.hotWaterTank[%s].values.heatRequestTop",
         value_type=str,
         human_readable=HotWaterTankHeatRequest,
     )
@@ -192,65 +192,129 @@ class HeatPump(Enum):
     """The heat pump endpoint settings."""
 
     NAME = EndpointProperties(
-        f"{PAYLOAD_PREFIX}.heatpump[%s].param.name",
+        f"{PAYLOAD_PREFIX}.sParam.heatpump[%s].param.name",
         value_type=str,
     )
     STATE = EndpointProperties(
-        f"{PAYLOAD_PREFIX}.heatpump[%s].values.heatpumpState",
+        f"{PAYLOAD_PREFIX}.sParam.heatpump[%s].values.heatpumpState",
         value_type=int,
         read_only=False,
         human_readable=HeatPumpState,
     )
     OPERATING_MODE = EndpointProperties(
-        f"{PAYLOAD_PREFIX}.heatpump[%s].param.operatingMode",
+        f"{PAYLOAD_PREFIX}.sParam.heatpump[%s].param.operatingMode",
         value_type=int,
         read_only=False,
         human_readable=HeatPumpOperatingMode,
     )
     CIRCULATION_PUMP = EndpointProperties(
-        f"{PAYLOAD_PREFIX}.heatpump[%s].CircPump.values.setValueScaled",
+        f"{PAYLOAD_PREFIX}.sParam.heatpump[%s].CircPump.values.setValueScaled",
         value_type=float,
     )
     INFLOW_TEMPERATURE = EndpointProperties(
-        f"{PAYLOAD_PREFIX}.heatpump[%s].TempHeatFlow.values.actValue",
+        f"{PAYLOAD_PREFIX}.sParam.heatpump[%s].TempHeatFlow.values.actValue",
         value_type=float,
     )
     REFLUX_TEMPERATURE = EndpointProperties(
-        f"{PAYLOAD_PREFIX}.heatpump[%s].TempHeatReflux.values.actValue",
+        f"{PAYLOAD_PREFIX}.sParam.heatpump[%s].TempHeatReflux.values.actValue",
         value_type=float,
     )
     SOURCE_INPUT_TEMPERATURE = EndpointProperties(
-        f"{PAYLOAD_PREFIX}.heatpump[%s].TempSourceIn.values.actValue",
+        f"{PAYLOAD_PREFIX}.sParam.heatpump[%s].TempSourceIn.values.actValue",
         value_type=float,
     )
     SOURCE_OUTPUT_TEMPERATURE = EndpointProperties(
-        f"{PAYLOAD_PREFIX}.heatpump[%s].TempSourceOut.values.actValue",
+        f"{PAYLOAD_PREFIX}.sParam.heatpump[%s].TempSourceOut.values.actValue",
         value_type=float,
     )
     COMPRESSOR_INPUT_TEMPERATURE = EndpointProperties(
-        f"{PAYLOAD_PREFIX}.heatpump[%s].TempCompressorIn.values.actValue",
+        f"{PAYLOAD_PREFIX}.sParam.heatpump[%s].TempCompressorIn.values.actValue",
         value_type=float,
     )
     COMPRESSOR_OUTPUT_TEMPERATURE = EndpointProperties(
-        f"{PAYLOAD_PREFIX}.heatpump[%s].TempCompressorOut.values.actValue",
+        f"{PAYLOAD_PREFIX}.sParam.heatpump[%s].TempCompressorOut.values.actValue",
         value_type=float,
     )
     COMPRESSOR = EndpointProperties(
-        f"{PAYLOAD_PREFIX}.heatpump[%s].Compressor.values.setValueScaled",
+        f"{PAYLOAD_PREFIX}.sParam.heatpump[%s].Compressor.values.setValueScaled",
         value_type=float,
     )
     HIGH_PRESSURE = EndpointProperties(
-        f"{PAYLOAD_PREFIX}.heatpump[%s].HighPressure.values.actValue",
+        f"{PAYLOAD_PREFIX}.sParam.heatpump[%s].HighPressure.values.actValue",
         value_type=float,
     )
     LOW_PRESSURE = EndpointProperties(
-        f"{PAYLOAD_PREFIX}.heatpump[%s].LowPressure.values.actValue",
+        f"{PAYLOAD_PREFIX}.sParam.heatpump[%s].LowPressure.values.actValue",
         value_type=float,
     )
     HEAT_REQUEST = EndpointProperties(
-        f"{PAYLOAD_PREFIX}.heatpump[%s].values.request",
+        f"{PAYLOAD_PREFIX}.sParam.heatpump[%s].values.request",
         value_type=str,
         human_readable=HeatPumpHeatRequest,
+    )
+    ELECTRICAL_ENERGY_METER = EndpointProperties(
+        f"{PAYLOAD_PREFIX}.sParam.heatpump[%s].ElectricEnergyMeter.values.power",
+        value_type=float,
+    )
+    HEAT_METER = EndpointProperties(
+        f"{PAYLOAD_PREFIX}.sParam.heatpump[%s].HeatMeter.values.power",
+        value_type=float,
+    )
+    HOT_WATER_METER = EndpointProperties(
+        f"{PAYLOAD_PREFIX}.sParam.heatpump[%s].HotWaterMeter.values.power",
+        value_type=float,
+    )
+    COP = EndpointProperties(
+        f"{PAYLOAD_PREFIX}.sParam.heatpump[%s].values.COP",
+        value_type=float,
+    )
+    CONSUMPTION_HEATING_ENERGY = EndpointProperties(
+        f"{PAYLOAD_PREFIX}.sStatisticalData.heatpump[%s].consumption.heating.energy",
+        value_type=float,
+    )
+    CONSUMPTION_HEATING_ELECTRICAL_ENERGY = EndpointProperties(
+        f"{PAYLOAD_PREFIX}.sStatisticalData.heatpump[%s].consumption.heating.electricalenergy",
+        value_type=float,
+    )
+    ENERGY_EFFICIENCY_RATIO_HEAT = EndpointProperties(
+        f"{PAYLOAD_PREFIX}.sStatisticalData.heatpump[%s].EnergyEfficiencyRatioHeat",
+        value_type=float,
+    )
+    CONSUMPTION_COOLING_ENERGY = EndpointProperties(
+        f"{PAYLOAD_PREFIX}.sStatisticalData.heatpump[%s].consumption.cooling.energy",
+        value_type=float,
+    )
+    CONSUMPTION_COOLING_ELECTRICAL_ENERGY = EndpointProperties(
+        f"{PAYLOAD_PREFIX}.sStatisticalData.heatpump[%s].consumption.cooling.electricalenergy",
+        value_type=float,
+    )
+    ENERGY_EFFICIENCY_RATIO_COOL = EndpointProperties(
+        f"{PAYLOAD_PREFIX}.sStatisticalData.heatpump[%s].EnergyEfficiencyRatioCool",
+        value_type=float,
+    )
+    CONSUMPTION_DOM_HOT_WATER_ENERGY = EndpointProperties(
+        f"{PAYLOAD_PREFIX}.sStatisticalData.heatpump[%s].consumption.domHotWater.energy",
+        value_type=float,
+    )
+    CONSUMPTION_DOM_HOT_WATER_ELECTRICAL_ENERGY = EndpointProperties(
+        f"{PAYLOAD_PREFIX}.sStatisticalData.heatpump[%s].consumption.domHotWater.electricalenergy",
+        value_type=float,
+    )
+    ENERGY_EFFICIENCY_RATIO_DOM_HOT_WATER = EndpointProperties(
+        f"{PAYLOAD_PREFIX}.sStatisticalData.heatpump[%s].EnergyEfficiencyRatioDomHotWater",
+        value_type=float,
+    )
+    CONSUMPTION_ENERGY = EndpointProperties(
+        f"{PAYLOAD_PREFIX}.sStatisticalData.heatpump[%s].consumption.energy",
+        value_type=float,
+    )
+    CONSUMPTION_ELECTRICAL_ENERGY = EndpointProperties(
+        f"{PAYLOAD_PREFIX}.sStatisticalData.heatpump[%s].consumption.electricalenergy",
+        value_type=float,
+    )
+    ENERGY_EFFICIENCY_RATIO = EndpointProperties(
+        f"{PAYLOAD_PREFIX}.sStatisticalData.heatpump[%s].EnergyEfficiencyRatio",
+        value_type=float,
     )
 
 
@@ -258,80 +322,80 @@ class HeatCircuit(Enum):
     """The heat circuit endpoint settings."""
 
     NAME = EndpointProperties(
-        f"{PAYLOAD_PREFIX}.heatCircuit[%s].param.name",
+        f"{PAYLOAD_PREFIX}.sParam.heatCircuit[%s].param.name",
         value_type=str,
     )
     HAS_ROOM_TEMPERATURE = EndpointProperties(
-        f"{PAYLOAD_PREFIX}.options.heatCircuit[%s].hasRoomTemp",
+        f"{PAYLOAD_PREFIX}.sParam.options.heatCircuit[%s].hasRoomTemp",
         value_type=str,
         human_readable=HeatCircuitHasRoomTemperature,
     )
     ROOM_TEMPERATURE = EndpointProperties(
-        f"{PAYLOAD_PREFIX}.heatCircuit[%s].tempRoom.values.actValue",
+        f"{PAYLOAD_PREFIX}.sParam.heatCircuit[%s].tempRoom.values.actValue",
         value_type=float,
     )
     HAS_ROOM_HUMIDITY = EndpointProperties(
-        f"{PAYLOAD_PREFIX}.options.heatCircuit[%s].hasRoomHumidity",
+        f"{PAYLOAD_PREFIX}.sParam.options.heatCircuit[%s].hasRoomHumidity",
         value_type=str,
         human_readable=HeatCircuitHasRoomHumidity,
     )
     ROOM_HUMIDITY = EndpointProperties(
-        f"{PAYLOAD_PREFIX}.heatCircuit[%s].humidityRoom.values.actValue",
+        f"{PAYLOAD_PREFIX}.sParam.heatCircuit[%s].humidityRoom.values.actValue",
         value_type=float,
     )
     DEW_POINT = EndpointProperties(
-        f"{PAYLOAD_PREFIX}.heatCircuit[%s].dewPoint.values.actValue",
+        f"{PAYLOAD_PREFIX}.sParam.heatCircuit[%s].dewPoint.values.actValue",
         value_type=float,
     )
     TEMPERATURE = EndpointProperties(
-        f"{PAYLOAD_PREFIX}.heatCircuit[%s].values.setValue",
+        f"{PAYLOAD_PREFIX}.sParam.heatCircuit[%s].values.setValue",
         value_type=float,
     )
     DAY_TEMPERATURE = EndpointProperties(
-        f"{PAYLOAD_PREFIX}.heatCircuit[%s].param.normalSetTemp",
+        f"{PAYLOAD_PREFIX}.sParam.heatCircuit[%s].param.normalSetTemp",
         value_type=float,
         read_only=False,
     )
     DAY_TEMPERATURE_THRESHOLD = EndpointProperties(
-        f"{PAYLOAD_PREFIX}.heatCircuit[%s].param.thresholdDayTemp.value",
+        f"{PAYLOAD_PREFIX}.sParam.heatCircuit[%s].param.thresholdDayTemp.value",
         value_type=float,
     )
     NIGHT_TEMPERATURE = EndpointProperties(
-        f"{PAYLOAD_PREFIX}.heatCircuit[%s].param.reducedSetTemp",
+        f"{PAYLOAD_PREFIX}.sParam.heatCircuit[%s].param.reducedSetTemp",
         value_type=float,
         read_only=False,
     )
     NIGHT_TEMPERATURE_THRESHOLD = EndpointProperties(
-        f"{PAYLOAD_PREFIX}.heatCircuit[%s].param.thresholdNightTemp.value",
+        f"{PAYLOAD_PREFIX}.sParam.heatCircuit[%s].param.thresholdNightTemp.value",
         value_type=float,
     )
     HOLIDAY_TEMPERATURE = EndpointProperties(
-        f"{PAYLOAD_PREFIX}.heatCircuit[%s].param.holidaySetTemp",
+        f"{PAYLOAD_PREFIX}.sParam.heatCircuit[%s].param.holidaySetTemp",
         value_type=float,
         read_only=False,
     )
     TEMPERATURE_OFFSET = EndpointProperties(
-        f"{PAYLOAD_PREFIX}.heatCircuit[%s].param.offsetRoomTemp",
+        f"{PAYLOAD_PREFIX}.sParam.heatCircuit[%s].param.offsetRoomTemp",
         value_type=float,
         read_only=False,
     )
     HEAT_REQUEST = EndpointProperties(
-        f"{PAYLOAD_PREFIX}.heatCircuit[%s].values.heatRequest",
+        f"{PAYLOAD_PREFIX}.sParam.heatCircuit[%s].values.heatRequest",
         value_type=str,
         human_readable=HeatCircuitHeatRequest,
     )
     EXTERNAL_COOL_REQUEST = EndpointProperties(
-        f"{PAYLOAD_PREFIX}.heatCircuit[%s].param.external.coolRequest",
+        f"{PAYLOAD_PREFIX}.sParam.heatCircuit[%s].param.external.coolRequest",
         value_type=str,
         human_readable=HeatCircuitExternalCoolRequest,
     )
     EXTERNAL_HEAT_REQUEST = EndpointProperties(
-        f"{PAYLOAD_PREFIX}.heatCircuit[%s].param.external.heatRequest",
+        f"{PAYLOAD_PREFIX}.sParam.heatCircuit[%s].param.external.heatRequest",
         value_type=str,
         human_readable=HeatCircuitExternalHeatRequest,
     )
     OPERATING_MODE = EndpointProperties(
-        f"{PAYLOAD_PREFIX}.heatCircuit[%s].param.operatingMode",
+        f"{PAYLOAD_PREFIX}.sParam.heatCircuit[%s].param.operatingMode",
         value_type=int,
         read_only=False,
         human_readable=HeatCircuitOperatingMode,