+++ CHANGELOG.md
@@ -7,6 +7,14 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+## [1.15.0] - 2025-09-14
+
+### Added
+
+- Add endpoints to read the energy management data from the API
+
+Add endpoints to read the energy management data from the API
+
 ## [1.14.0] - 2025-08-27
 
 ### Added